// libs/interfaces.ts
export interface Data {
  destinations: [
    {
      name: string;
      images: {
        webp: string;
      };
      description: string;
      distance: string;
      travel: string;
    },
    {
      name: string;
      images: {
        webp: string;
      };
      description: string;
      distance: string;
      travel: string;
    },
    {
      name: string;
      images: {
        webp: string;
      };
      description: string;
      distance: string;
      travel: string;
    },
    {
      name: string;
      images: {
        webp: string;
      };
      description: string;
      distance: string;
      travel: string;
    }
  ];
  crew: [
    {
      name: string;
      images: {
        webp: string;
      };
      role: string;
      bio: string;
    },
    {
      name: string;
      images: {
        webp: string;
      };
      role: string;
      bio: string;
    },
    {
      name: string;
      images: {
        webp: string;
      };
      role: string;
      bio: string;
    },
    {
      name: string;
      images: {
        webp: string;
      };
      role: string;
      bio: string;
    }
  ];
  technology: [
    {
      name: string;
      images: {
        portrait: string;
        landscape: string;
      };
      description: string;
    },
    {
      name: string;
      images: {
        portrait: string;
        landscape: string;
      };
      description: string;
    },
    {
      name: string;
      images: {
        portrait: string;
        landscape: string;
      };
      description: string;
    }
  ];
}